fix. 컴포넌트 폴더구조 정리
