+++ src/page/form/component/index.tsx
@@ -1,0 +1,11 @@
+import LoginForm from "./organism/Form";
+
+const FormTemplate = () => {
+  return (
+    <>
+      <LoginForm />
+    </>
+  );
+};
+
+export default FormTemplate;
