+++ src/page/form/component/organism/Link.tsx
@@ -1,0 +1,3 @@
+const Link = () => {
+  return <></>;
+};
